Refactored SQL Server access
In the next couple of commits, when I implement actual unit tests, it'll
be useful to have a dependency-injected data source for mocking.
