+++ EVE-Val SQL to YAML Tool/SQL To YAML CLI/Program.cs
@@ -18,8 +18,8 @@ namespace SQL_To_YAML_CLI
                 return;
             }
             StreamWriter w = new StreamWriter(options.outputFile);
-            SQLToYAML sty = new SQLToYAML(options.table, w, options.notificationPercent);
-            sty.ConnectionString = options.connectionString;
+            DataLayer d = new DataLayer(options.connectionString);
+            SQLToYAML sty = new SQLToYAML(options.table, w, d, options.notificationPercent);
             sty.MadeProgress += new MadeProgressEventHandler(sty_MadeProgress);
             sty.ExtractionFinished += new ExtractionFinishedEventHandler(sty_ExtractionFinished);
             sty.StartProcessing();
