+++ EVE-Val SQL to YAML Tool/SQL To YAML CLI/SQL To YAML CLI.csproj.user
@@ -11,11 +11,11 @@
     <VerifyUploadedFiles>false</VerifyUploadedFiles>
   </PropertyGroup>
   <PropertyGroup Condition="'$(Configuration)|$(Platform)' == 'Debug|x86'">
-    <StartArguments>-o silas.yaml</StartArguments>
+    <StartArguments>-t invCategories -o silas.yaml</StartArguments>
     <StartWorkingDirectory>C:\</StartWorkingDirectory>
   </PropertyGroup>
   <PropertyGroup Condition="'$(Configuration)|$(Platform)' == 'Release|x86'">
-    <StartArguments>-o silas.yaml</StartArguments>
+    <StartArguments>-t invCategories -o silas.yaml</StartArguments>
     <StartWorkingDirectory>C:\</StartWorkingDirectory>
   </PropertyGroup>
 </Project>
