+++ SQLToYAML/ExtractorLib.csproj
@@ -40,6 +40,7 @@
     <Reference Include="System.Xml" />
   </ItemGroup>
   <ItemGroup>
+    <Compile Include="DataLayer.cs" />
     <Compile Include="SQLToYAML.cs" />
     <Compile Include="Properties\AssemblyInfo.cs" />
     <Compile Include="YAMLConversion.cs" />
